[cleanup]: size_t -> uint64_t

+++ include/crypto/chacha20.h
@@ -4,4 +4,4 @@
 
 void chacha20_encrypt(const uint8_t key[32], const uint8_t nonce[12],
                       uint32_t counter, const uint8_t *in, uint8_t *out,
-                      size_t len);
+                      uint64_t len);
